Delete domain layer dependency
