+++ data/src/main/java/es/architectcoders/data/repository/ApodRepository.kt
@@ -1,11 +1,12 @@
 package es.architectcoders.data.repository
 
+import es.architectcoders.data.mappers.toDomain
 import es.architectcoders.data.source.local.ApodLocalDataSource
 import es.architectcoders.data.source.network.ApodRemoteDataSource
-import es.architectcoders.framework.model.toError
 import es.architectcoders.domain.model.Apod
 import es.architectcoders.domain.model.Error
 import es.architectcoders.framework.model.ApodResponse
+import es.architectcoders.framework.model.toError
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
@@ -23,7 +24,7 @@ class ApodRepository @Inject constructor(
         }
         null // return null if no error
     } catch (exception: Exception) {
-        exception.toError()
+        exception.toError().toDomain()
     }
 
     fun getApods(): Flow<List<Apod>> {
@@ -34,6 +35,6 @@ class ApodRepository @Inject constructor(
         apodLocalDataSource.saveApodAsFavourite(apod)
         null // return null if no error
     } catch (exception: Exception) {
-        exception.toError()
+        exception.toError().toDomain()
     }
 }
