+++ framework/build.gradle.kts
@@ -75,8 +75,6 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-
-    implementation(project(mapOf("path" to ":domain")))
 }
 
 // Allow references to generated code
